+++ src/app/app.component.html
@@ -1,139 +1,1 @@
-<div class="header">
-  <img src="https://github.com/VinyChagas/SiteSCS/blob/2f8b6b9c91d67478583edcecfc0318fbbc8e2656/images/default-logo.png?raw=true" alt="Logo" class="logo">
-  <button class="import-button">Importar Cartão CNPJ</button>
-</div>
-
-<div class="form-container">
-  <h2>Cadastro de Empresa</h2>
-  <form #empresaForm="ngForm" (ngSubmit)="salvar()">
-    <label>Razão Social: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.razao_social" name="razao_social" required maxlength="255" (blur)="validarRazaoSocial()" [ngClass]="{'ng-invalid': erros.razao_social}">
-    <span *ngIf="erros.razao_social" class="error-message">{{ erros.razao_social }}</span><br>
-
-    <label>Nome Fantasia:</label>
-    <input type="text" [(ngModel)]="empresa.nome_fantasia" name="nome_fantasia" maxlength="255" (focus)="preencherNomeFantasia()" [ngClass]="{'ng-invalid': erros.nome_fantasia}">
-    <span *ngIf="erros.nome_fantasia" class="error-message">{{ erros.nome_fantasia }}</span><br>
-
-    <label>CNPJ: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.cnpj" name="cnpj" required
-           (keypress)="permitirSomenteNumeros($event)"
-           (input)="aplicarMascaraCNPJ($event)"
-           maxlength="18"
-           placeholder="00.000.000/0000-00"
-           [ngClass]="{'ng-invalid': erros.cnpj}">
-    <span *ngIf="erros.cnpj" class="error-message">{{ erros.cnpj }}</span><br>
-
-    <label>Inscrição Estadual:</label>
-    <input type="text" [(ngModel)]="empresa.inscricao_estadual" name="inscricao_estadual"
-           (input)="aplicarMascaraInscricaoEstadual($event)" 
-           maxlength="10"
-           placeholder="00.000.000-0"
-           [ngClass]="{'ng-invalid': erros.inscricao_estadual}">
-    <span *ngIf="erros.inscricao_estadual" class="error-message">{{ erros.inscricao_estadual }}</span><br>
-
-    <label>Inscrição Municipal:</label>
-    <input type="text" [(ngModel)]="empresa.inscricao_municipal" name="inscricao_municipal"
-           (input)="aplicarMascaraInscricaoMunicipal($event)"
-           maxlength="9"
-           placeholder="00.00000-0"
-           [ngClass]="{'ng-invalid': erros.inscricao_municipal}">
-    <span *ngIf="erros.inscricao_municipal" class="error-message">{{ erros.inscricao_municipal }}</span><br>
-
-    <label>CEP: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.cep" name="cep" required
-           (input)="aplicarMascaraCEP($event)"
-           maxlength="9"
-           placeholder="00000-000"
-           [ngClass]="{'ng-invalid': erros.cep}">
-    <span *ngIf="erros.cep" class="error-message">{{ erros.cep }}</span><br>
-
-    <label>Estado: <span class="required">*</span></label>
-    <select [(ngModel)]="empresa.estado" name="estado" required [ngClass]="{'ng-invalid': erros.estado}">
-      <option value="" disabled selected>Selecione o estado</option>
-      <option *ngFor="let estado of estados" [value]="estado">{{ estado }}</option>
-    </select>
-    <span *ngIf="erros.estado" class="error-message">{{ erros.estado }}</span><br>
-
-    <label>Cidade: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.cidade" name="cidade" required
-           maxlength="100"
-           placeholder="Informe a cidade"
-           [ngClass]="{'ng-invalid': erros.cidade}">
-    <span *ngIf="erros.cidade" class="error-message">{{ erros.cidade }}</span><br>
-
-    <label>Bairro: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.bairro" name="bairro" required
-           maxlength="100"
-           placeholder="Informe o bairro"
-           [ngClass]="{'ng-invalid': erros.bairro}">
-    <span *ngIf="erros.bairro" class="error-message">{{ erros.bairro }}</span><br>
-
-    <label>Logradouro: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.logradouro" name="logradouro" required
-           maxlength="255"
-           placeholder="Informe o logradouro"
-           [ngClass]="{'ng-invalid': erros.logradouro}">
-    <span *ngIf="erros.logradouro" class="error-message">{{ erros.logradouro }}</span><br>
-
-    <label>Número: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.numero" name="numero" required
-           maxlength="7"
-           placeholder="Informe o número"
-           [ngClass]="{'ng-invalid': erros.numero}">
-    <span *ngIf="erros.numero" class="error-message">{{ erros.numero }}</span><br>
-
-    <label>Complemento:</label>
-    <input type="text" [(ngModel)]="empresa.complemento" name="complemento"
-           maxlength="100"
-           placeholder="Apto, bloco, sala..."
-           [ngClass]="{'ng-invalid': erros.complemento}">
-    <span *ngIf="erros.complemento" class="error-message">{{ erros.complemento }}</span><br>
-
-    <label>Telefone:</label>
-    <input type="text" [(ngModel)]="empresa.telefone" name="telefone"
-           (input)="aplicarMascaraTelefone($event)"
-           maxlength="15"
-           placeholder="(99) 99999-9999"
-           [ngClass]="{'ng-invalid': erros.telefone}">
-    <span *ngIf="erros.telefone" class="error-message">{{ erros.telefone }}</span><br>
-
-    <label>Email:</label>
-    <input type="email" [(ngModel)]="empresa.email" name="email" maxlength="100"
-           placeholder="exemplo@email.com"
-           [ngClass]="{'ng-invalid': erros.email}"
-           title="{{ erros.email }}">
-    <span *ngIf="erros.email" class="error-message">{{ erros.email }}</span><br>
-
-    <label>Site:</label>
-    <input type="text" [(ngModel)]="empresa.site" name="site"
-           placeholder="https://www.exemplo.com"
-           [ngClass]="{'ng-invalid': erros.site}"
-           title="{{ erros.site }}">
-    <span *ngIf="erros.site" class="error-message">{{ erros.site }}</span><br>
-
-    <label>Regime Tributário: <span class="required">*</span></label>
-    <select [(ngModel)]="empresa.regime_tributario" name="regime_tributario" required [ngClass]="{'ng-invalid': erros.regime_tributario}">
-      <option value="" disabled selected>Selecione o regime tributário</option>
-      <option *ngFor="let regime of regimesTributarios" [value]="regime">{{ regime }}</option>
-    </select>
-    <span *ngIf="erros.regime_tributario" class="error-message">{{ erros.regime_tributario }}</span><br>
-
-    <label>CNAE Principal: <span class="required">*</span></label>
-    <input type="text" [(ngModel)]="empresa.cnae_principal" name="cnae_principal" required
-           (input)="aplicarMascaraCNAE($event)"
-           maxlength="9"
-           placeholder="0000-0/00"
-           [ngClass]="{'ng-invalid': erros.cnae_principal}">
-    <span *ngIf="erros.cnae_principal" class="error-message">{{ erros.cnae_principal }}</span><br>
-
-    <label>Data de Abertura: <span class="required">*</span></label>
-    <input type="date" [(ngModel)]="empresa.data_abertura" name="data_abertura" required
-           [max]="today"
-           [ngClass]="{'ng-invalid': erros.data_abertura}">
-    <span *ngIf="erros.data_abertura" class="error-message">{{ erros.data_abertura }}</span><br>
-
-    <br>
-    <button type="submit">Salvar</button>
-    <button type="button" (click)="excluir()">Excluir</button>
-  </form>
-</div>
+<router-outlet></router-outlet>
