+++ src/app/form/form.component.css
@@ -1,0 +1,100 @@
+form {
+    max-width: 500px;
+    margin: 0 auto; /* Centraliza o card horizontalmente */
+    padding: 20px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    background-color: #f9f9f9;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+label {
+    display: block;
+    margin-top: 10px;
+}
+
+input, select {
+    width: 100%;
+    padding: 8px;
+    margin-top: 5px;
+    margin-bottom: 15px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    box-sizing: border-box;
+}
+
+button {
+    margin-top: 15px;
+    padding: 10px 15px;
+}
+
+button + button {
+    margin-left: 20px; /* Adiciona espaçamento de 20px entre os botões */
+    margin-top: 10px; /* Adiciona espaçamento de 10px acima do segundo botão */
+}
+
+h2 {
+    margin-top: 15px;
+    padding: 10px 15px;
+    text-align: center; /* Centraliza o texto */
+}
+
+/* Estilo para campos com erro */
+input.ng-invalid.ng-dirty {
+    border: 1px solid red;
+    background-color: #ffe6e6;
+}
+
+/* Estilo para mensagens de erro */
+span.error-message {
+    color: red;
+    font-size: 12px;
+    margin-top: 5px;
+    display: block;
+}
+
+.logo {
+    height: 60px; /* Aumenta a altura da logo em 50% (de 40px para 60px) */
+    width: auto; /* Mantém a proporção da largura */
+}
+
+.header {
+    display: flex;
+    justify-content: space-between; /* Espaça os itens no cabeçalho */
+    align-items: center;
+    padding: 10px 20px;
+    background-color: #fff;
+    border-bottom: 1px solid #ccc;
+}
+
+.nav-links {
+    display: flex;
+    gap: 30px; /* Aumenta o espaçamento entre os links de 20px para 30px */
+}
+
+.nav-links a {
+    text-decoration: none; /* Remove o sublinhado dos links */
+    color: #333; /* Cor do texto */
+    font-weight: bold; /* Destaca os links */
+    transition: color 0.3s; /* Transição suave para hover */
+}
+
+.nav-links a:hover {
+    color: #007bff; /* Cor ao passar o mouse */
+}
+
+.import-button {
+    padding: 10px 20px; /* Espaçamento interno */
+    font-size: 14px; /* Tamanho da fonte */
+    color: #fff; /* Cor do texto */
+    background-color: #007bff; /* Cor de fundo */
+    border: none; /* Remove a borda padrão */
+    border-radius: 4px; /* Cantos arredondados */
+    cursor: pointer; /* Altera o cursor ao passar sobre o botão */
+    transition: background-color 0.3s; /* Transição suave para hover */
+    margin-top: -5px; /* Move o botão 5px para cima */
+}
+
+.import-button:hover {
+    background-color: #0056b3; /* Cor ao passar o mouse */
+}
