+++ src/app/login/login.component.html
@@ -1,0 +1,24 @@
+<div class="login-page">
+  <div class="login-container">
+    <div class="login-image">
+      <!-- Espaço reservado para a imagem -->
+    </div>
+    <div class="login-form">
+      <h2>Bem-vindo</h2>
+      <p>Faça login para acessar sua conta</p>
+      <form>
+        <label for="email">Email:</label>
+        <input type="email" id="email" name="email" placeholder="Digite seu email" required>
+        <label for="password">Senha:</label>
+        <input type="password" id="password" name="password" placeholder="Digite sua senha" required>
+        <button type="button">Entrar</button>
+        <div class="login-options">
+          <p>Ou faça login com:</p>
+          <button type="button" class="social-button">Google</button>
+          <button type="button" class="social-button">Facebook</button>
+          <button type="button" class="social-button">Twitter</button>
+        </div>
+      </form>
+    </div>
+  </div>
+</div>
